inverse_diffusion/requirements.txt
accelerate==0.27.2
apex==0.9.10dev
deeplake==3.9.4
denoising_diffusion_pytorch==1.10.12
einops==0.8.0
ema_pytorch==0.4.2
fld==1.0.0
gdown==5.2.0
huggingface_hub==0.22.2
matplotlib==3.8.0
numpy==1.26.4
packaging==24.0
peft==0.10.0
Pillow==10.3.0
pytorch_fid==0.3.0
scheduling_utils==0.2.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.6
